separated the package dns
